+++ react-ufo/atlaskit/src/load-hold/UFOLoadHold.tsx
@@ -24,7 +24,8 @@ const useLayoutEffectSAFE = typeof window === 'undefined' ? useEffect : useLayou
  *   return (
  *     <>
  *       <Skeleton />
- *       <UFOLoadHold name="card" />
+ *       <UFOLoadHold name="card">
+ *     </UFOLoadHold>
  *   )
  * }
  * ```
@@ -52,6 +53,8 @@ type Props = {
 	name: string;
 	hold?: boolean;
 	children?: ReactNode;
+
+	// TODO: implement Experimental UFO Holds functionality as per https://product-fabric.atlassian.net/browse/AFO-3080
 	experimental?: boolean;
 };
 
@@ -60,10 +63,7 @@ export default function UFOLoadHold({ children, name, hold = true, experimental 
 	const context = useContext(UFOInteractionContext);
 
 	useLayoutEffectSAFE(() => {
-		if (hold && context != null) {
-			if (experimental && context.holdExperimental) {
-				return context.holdExperimental(name);
-			}
+		if (hold && !experimental && context != null) {
 			return context.hold(name);
 		}
 	}, [hold, context, name]);