+++ react-ufo/atlaskit/src/segment/segment.tsx
@@ -114,14 +114,14 @@ export default function UFOSegment({ name: segmentName, children }: Props) {
 			this: EnhancedUFOInteractionContextType,
 			labelStack: LabelStack,
 			name: string,
-			experimental = false,
-		) {
+			// eslint-disable-next-line @typescript-eslint/no-invalid-void-type
+		): (() => void) | void {
 			if (interactionId.current != null) {
 				if (parentContext) {
-					return parentContext._internalHold(labelStack, name, experimental);
+					return parentContext._internalHold(labelStack, name);
 				} else {
 					const capturedInteractionId = interactionId.current;
-					const disposeHold = addHold(interactionId.current, labelStack, name, experimental);
+					const disposeHold = addHold(interactionId.current, labelStack, name);
 					return () => {
 						if (capturedInteractionId === interactionId.current) {
 							disposeHold();
@@ -137,7 +137,8 @@ export default function UFOSegment({ name: segmentName, children }: Props) {
 			id: string,
 			name: string,
 			remove: boolean,
-		) {
+			// eslint-disable-next-line @typescript-eslint/no-invalid-void-type
+		): (() => void) | void {
 			if (interactionId.current != null) {
 				if (parentContext) {
 					parentContext._internalHoldByID(labelStack, name, id, remove);
@@ -162,11 +163,12 @@ export default function UFOSegment({ name: segmentName, children }: Props) {
 
 		return {
 			labelStack,
-			hold(this: EnhancedUFOInteractionContextType, name: string | undefined = 'unknown') {
+			hold(
+				this: EnhancedUFOInteractionContextType,
+				name: string | undefined = 'unknown',
+				// eslint-disable-next-line @typescript-eslint/no-invalid-void-type
+			): void | (() => void) {
 				return this._internalHold(this.labelStack, name);
-			},
-			holdExperimental(this: EnhancedUFOInteractionContextType, name: string = 'unknown') {
-				return this._internalHold(this.labelStack, name, true);
 			},
 			addHoldByID(
 				this: EnhancedUFOInteractionContextType,