+++ editor/editor-plugins/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugins
 
+## 5.9.24
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 5.9.23
 
 ### Patch Changes